# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Voisso API",
    "description": "Voisso marketing automation platform API",
    "version": "0.1.0"
  },
  "paths": {
    "/brands/{brand_id}/campaigns": {
      "get": {
        "tags": [
          "Brand Campaigns"
        ],
        "summary": "List",
        "operationId": "brand_campaigns_list",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CampaignState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignResponse"
                  },
                  "title": "Response Brand Campaigns List"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/brands/{brand_id}/campaigns/{campaign_id}": {
      "delete": {
        "tags": [
          "Brand Campaigns"
        ],
        "summary": "Delete Campaign",
        "operationId": "brand_campaigns_delete_campaign",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/brands": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "Search",
        "operationId": "brands_search",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BrandResponse"
                  },
                  "title": "Response Brands Search"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "summary": "Create",
        "operationId": "brands_create",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/brands/{brand_id}": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "Get",
        "operationId": "brands_get",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "summary": "Update",
        "operationId": "brands_update",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_brands_update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "summary": "Delete",
        "operationId": "brands_delete",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/brand-generation": {
      "post": {
        "tags": [
          "brand-generation"
        ],
        "summary": "Start",
        "operationId": "brand_generation_start",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandGenerationJobCreateRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandGenerationJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/brand-generation/{job_id}": {
      "get": {
        "tags": [
          "brand-generation"
        ],
        "summary": "Get",
        "operationId": "brand_generation_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandGenerationJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/brand-generation/{job_id}/accept": {
      "post": {
        "tags": [
          "brand-generation"
        ],
        "summary": "Accept",
        "operationId": "brand_generation_accept",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandGenerationJobAcceptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_brands_update": {
        "properties": {
          "request_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Data"
          },
          "logo_file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo File"
          }
        },
        "type": "object",
        "title": "Body_brands_update"
      },
      "BrandArchetypeName": {
        "type": "string",
        "enum": [
          "INNOCENT",
          "EVERYMAN",
          "HERO",
          "OUTLAW",
          "EXPLORER",
          "CREATOR",
          "RULER",
          "MAGICIAN",
          "LOVER",
          "CAREGIVER",
          "JESTER",
          "SAGE"
        ],
        "title": "BrandArchetypeName"
      },
      "BrandAudience": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "age_range": {
            "$ref": "#/components/schemas/BrandAudienceAgeRange",
            "default": "ANY"
          },
          "gender": {
            "$ref": "#/components/schemas/BrandAudienceGender",
            "default": "ANY"
          },
          "income_range": {
            "$ref": "#/components/schemas/BrandAudienceIncomeRange",
            "default": "ANY"
          },
          "pain_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pain Points"
          },
          "objections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Objections"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ContentChannelName"
            },
            "type": "array",
            "title": "Channels"
          }
        },
        "type": "object",
        "title": "BrandAudience"
      },
...
      "BrandColor": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "hex_value": {
            "type": "string",
            "title": "Hex Value"
          }
        },
        "type": "object",
        "required": [
          "name",
          "hex_value"
        ],
        "title": "BrandColor"
      },
      "BrandCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "data": {
            "$ref": "#/components/schemas/BrandData"
          }
        },
        "type": "object",
        "required": [
          "name",
          "data"
        ],
        "title": "BrandCreateRequest"
      },
...
      "BrandData": {
        "properties": {
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "media_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Media Urls"
          },
          "colors": {
            "items": {
              "$ref": "#/components/schemas/BrandColor"
            },
            "type": "array",
            "title": "Colors"
          },
          "brand_mission": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Mission"
          },
          "archetype": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandArchetypeName"
              },
              {
                "type": "null"
              }
            ]
          },
          "locale": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\w{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "audiences": {
            "items": {
              "$ref": "#/components/schemas/BrandAudience"
            },
            "type": "array",
            "title": "Audiences"
          },
          "content_pillars": {
            "items": {
              "$ref": "#/components/schemas/ContentPillar"
            },
            "type": "array",
            "title": "Content Pillars"
          },
          "tone_of_voice": {
            "$ref": "#/components/schemas/BrandToneOfVoice"
          },
          "positioning": {
            "$ref": "#/components/schemas/PositioningBrandData"
          }
        },
        "type": "object",
        "title": "BrandData"
      },
      "BrandGenerationJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "website_url": {
            "type": "string",
            "title": "Website Url"
          },
          "extra_routes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extra Routes",
            "default": [
              "/about",
              "/help"
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandGenerationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "status",
          "website_url",
          "created_at",
          "updated_at"
        ],
        "title": "BrandGenerationJob"
      },
      "BrandGenerationJobAcceptRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the brand"
          },
          "data": {
            "$ref": "#/components/schemas/BrandData"
          }
        },
        "type": "object",
        "required": [
          "name",
          "data"
        ],
        "title": "BrandGenerationJobAcceptRequest"
      },
      "BrandGenerationJobCreateRequestBody": {
        "properties": {
          "website_url": {
            "type": "string",
            "title": "Website Url",
            "description": "The URL of the website to extract the brand from"
          }
        },
        "type": "object",
        "required": [
          "website_url"
        ],
        "title": "BrandGenerationJobCreateRequestBody"
      },
      "BrandGenerationJobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandGenerationJobResultResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "BrandGenerationJobResponse"
      },
      "BrandGenerationJobResultResponse": {
        "properties": {
          "brand_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandCreateRequestResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "scraper_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScrapeResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "BrandGenerationJobResultResponse"
      },
      "BrandGenerationResult": {
        "properties": {
          "brand_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandCreateRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "scraper_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScrapeResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "BrandGenerationResult"
      },
      "BrandResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandData"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "BrandResponse"
      },
...
      "CampaignData": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "goal": {
            "$ref": "#/components/schemas/ContentPillarBusinessGoal",
            "default": "INCREASE_CONVERSION"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "target_audience_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Audience Ids"
          },
          "content_pillar_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Pillar Ids"
          },
          "content_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Ids"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ContentChannelName"
            },
            "type": "array",
            "title": "Channels"
          },
          "media_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Media Urls"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date"
        ],
        "title": "CampaignData"
      },
...
      "CampaignResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "brand_id": {
            "type": "string",
            "title": "Brand Id"
          },
          "state": {
            "$ref": "#/components/schemas/CampaignState"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CampaignData"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "brand_id",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "CampaignResponse"
      },
      "CampaignState": {
        "type": "string",
        "enum": [
          "DRAFT",
          "LIVE",
          "PAUSED",
          "ENDED"
        ],
        "title": "CampaignState"
      },
...
      "ContentChannelName": {
        "type": "string",
        "enum": [
          "INSTAGRAM",
          "LINKEDIN"
        ],
        "title": "ContentChannelName"
      },
...
      "ContentPillarBusinessGoal": {
        "type": "string",
        "enum": [
          "DRIVE_ENGAGEMENT",
          "INCREASE_CONVERSION",
          "BUILD_TRUST",
          "GENERATE_LEADS",
          "RETENTION"
        ],
        "title": "ContentPillarBusinessGoal"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "cancelled",
          "failed",
          "accepted"
        ],
        "title": "JobStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}